ci(cache): include Python version in cache key for venv
The venv contains a symlink to the "parent" Python executables, whose
location changes on version changes, breaking the symlinks.

Include Python version in the cache key to prevent that in common cases.

This does not prevent breakage if the executable paths change without
the version changing, too. Accept that risk.