refactor(retry): simplify, err on safe side
Removes need to maintain info whether we want to treat error responses
as requiring re-login. We also now login when it's not detected that we
don't need to, and re-raise the original error on failures instead of
crafting our own.

A undesirable side effect of always doing raise_on_status is that the
error response for operations raising when not logged in is not recorded
in the cassette, making the error handling basically untested. But I
think this is a net positive given that drawback.