ci(release): do not require checks
We want to record changelog entries for commits that cause a CI failure
too. Removing the dependency on checks also enables to run the release
workflow in parallel with them.

If something fails with checks on merging the release triggering
release-please PR, then so be it and let the release proceed. That PR
should contain only unlinted markdown and release bumps in the first
place, which are reviewed in the release PR, so a failure is quite
unlikely there.